Provide dbus storage backend
This commit provides a working storage backend that allows clients to
connect transparently to a running hamster dbus service/objects.

Whilst this backend is mostly complient with the API defined in
``hamster_lib.storage`` we currently do not understand enough about dbus
related Exception/Error handling to implement all Exceptions defined.
As this backend is basicly just a serialization wrapper that calls the
service (which in turn then most likly uses a seperate backend to
persistently store the data) we still need to find a way to pass any
errors in the server on to the client end of things.

In order to test this backend, we make heacy use of ``python-dbusmock``.
To do so we use ``UnitTest`` style tests instead of our regular py.test
test and use py.test just as a testrunner.

Closes: DBUS-8