Add integration tests for dbus objects
After a conversation with Martin Pitt it became clear that there seems
to be no existing approach for proper unittests for dbus 'service
objects'. The recomended solution is to just write integration tests.
This is what this commit provides.

Besides moving ``hamster_dbus.objects`` related tests and fixtures to a
seperate directory we also provide two important new fixtures.
- ``private_session_bus`` takes an example from ``python-dbusmock`` in
providing a new private session bus in order to not pollute the regular
session bus.
- ``live_service`` launches a trivial service connected to our private
bus and exports the objects to be tested to it. Data stored by this
service is stored in an 'in-memory' sqlite database.

By using such a setup we are able to reuse most of our test that have
been removed in ``fa3babc342667a4d63f65b4281087849e2faf845``.

One unfortunate consequence of this integration test based solution is
that we can not actually reach full coverage for
``hamster_dbus.objects``, but this is more of a superficial issue.